refactor(frontend): config
