+++ frontend/src/config.js
@@ -1,8 +1,6 @@
-const networks = [
+export const networks = [
   {
     key: 'tequila-0004',
     lcd: 'https://tequila-fcd.terra.dev',
   },
 ];
-
-export default networks;
